+++ SharePointFramework/ProjectWebParts/src/components/ProjectInformation/index.tsx
@@ -56,6 +56,7 @@ export class ProjectInformation extends BaseWebPartComponent<
 
   public render() {
     if (this.state.hidden) return null
+
     return (
       <div className={styles.root}>
         <div className={styles.container}>