+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/index.tsx
@@ -25,7 +25,8 @@ import {
   ProjectPropertiesSection,
   RiskSection,
   StatusSection,
-  SummarySection
+  SummarySection,
+  TimelineSection,
 } from './Sections'
 import {
   IProjectStatusData,
@@ -69,11 +70,14 @@ export class ProjectStatus extends React.Component<IProjectStatusProps, IProject
           (report) => report.id === parseInt(selectedReportUrlParam, 10)
         )
       }
+      const newestReportId = data.reports.length > 0 ? data.reports[0].id : 0
+
       this.setState({
         data,
         selectedReport,
         sourceUrl: decodeURIComponent(sourceUrlParam || ''),
-        loading: false
+        loading: false,
+        newestReportId
       })
     } catch (error) {
       this.setState({ error, loading: false })
@@ -212,13 +216,13 @@ export class ProjectStatus extends React.Component<IProjectStatusProps, IProject
         disabled: true
       })
     }
-
     return (
       <CommandBar
         items={removeMenuBorder<IContextualMenuItem>(items)}
         farItems={removeMenuBorder<IContextualMenuItem>(farItems)}
       />
     )
+    
   }
 
   /**
@@ -246,7 +250,8 @@ export class ProjectStatus extends React.Component<IProjectStatusProps, IProject
       data: this.state.data,
       hubSiteUrl: this.props.hubSite.url,
       siteId: this.props.siteId,
-      webUrl: this.props.webUrl
+      webUrl: this.props.webUrl,
+      showLists: this.state.data.reports ? this.state.selectedReport.id === this.state.newestReportId : true
     }
     return baseProps
   }
@@ -309,6 +314,14 @@ export class ProjectStatus extends React.Component<IProjectStatusProps, IProject
           }
           case SectionType.ListSection: {
             return <ListSection {...baseProps} />
+          }
+          case SectionType.TimelineSection: {
+            return (
+              <TimelineSection
+                {...baseProps}
+                {...this.props}
+              />
+            )
           }
           default: {
             return null
@@ -478,7 +491,7 @@ export class ProjectStatus extends React.Component<IProjectStatusProps, IProject
           'PROJECT_STATUS',
           // eslint-disable-next-line quotes
           "Hidden eq false and Group ne 'Hidden'"
-        )
+        ),
       ])
       const sortedReports = reports
         .map((item) => item.setDefaultEditFormUrl(reportList.DefaultEditFormUrl))